NutriTrack/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
pymysql
ultralytics
xlsxwriter
# requests # implicitly used
# pillow # implicitly used
